requirements.txt
django==2.2.10
django-environ==0.4.5
djangorestframework==3.11.0
django-filter==2.0.0
django-url-filter==0.3.12
djangorestframework-jwt==1.11.0
ipdb==0.11
mypy==0.650
pytest==5.3.1
pytest-django==3.7.0
pytest-cov==2.6.1
django-cors-headers==3.2.0
requests==2.22.0
python-dotenv==0.10.3
djangorestframework-gis==0.15